requirements.txt
pyzmq==22.2.1
msgpack==1.0.2
playsound==1.3.0
argparse==1.4.0
torch==1.13.0
